CODE_MAPPER/requirements.txt
openai>=1.50.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.1.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
rich>=13.7.0
numpy>=1.26.0
tenacity>=8.2.0
PyYAML>=6.0.1
semgrep>=1.86.0
